v1.2: Implemented a management system for vector databases
The database after vectorization can be saved. There is no need to repeatedly input the saved CSV file. After selecting the CSV file, the hash value will be automatically calculated. If it matches the previously stored information, there is no need to repeat the embedding process. The preview, switching and deletion of the database are supported.